requirements.txt
djangorestframework
django
Pillow
python-dotenv
django-cors-headers